requirements.txt
pytest
pytest-playwright
allure-pytest